Update for consistent paths.
