+++ env.example
@@ -10,13 +10,20 @@ SECRET_KEY=your-secret-key-here-change-in-production
 # IMPORTANT: Generate a secure random key in production!
 JWT_SECRET_KEY=your-jwt-secret-key-here-change-in-production
 
-# Web UI Configuration
-WEB_PORT=8000
+# Server Configuration
+# Default host to bind to (0.0.0.0 for all interfaces, 127.0.0.1 for localhost only)
+DEFAULT_HOST=0.0.0.0
+# Default port to bind to (can be overridden with --port CLI flag)
+DEFAULT_PORT=8000
+
+# Legacy configuration (deprecated - use DEFAULT_PORT instead)
+# WEB_PORT=8000
 
 # MCP Server Configuration
 MCP_PATH=/mcp
 
-# Server URL (auto-configured based on WEB_PORT, but can override)
+# Server URL (auto-configured based on DEFAULT_HOST and DEFAULT_PORT, but can override)
+# Used for OAuth callbacks and external references
 # SERVER_URL=http://localhost:8000
 
 # Data Storage Configuration
